ComToTcpServer: Switch to new flag for formatted output
The original intention of the verbose flag was to enable debugger the debugger
traffic, but an emerged use case of the flag is to monitor device logging
while not broken into the debugger. This commits removed the verbose option and adds
an output option that cleanly outputs the traffic and a debug option intended for debug
prints.