UefiExt: Add Support to Print Cache Attr in PT Walk
This uses newly added Patina monitor commands to get
the MTRR value on x86 for a PA and the MAIR on ARM64
to parse the CacheIndex.

This is a backwards compatible change, if the monitor
command/output is not available, "Unknown" will be
printed instead of the cache type.

In addition, this updates the ARM64 !pt command to
read TTBR0_EL2 automatically instead of requiring it
to be passed in.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>